+++ db/seeds.rb
@@ -50,7 +50,7 @@ lot70 = FactoryBot.create(:lot, :paid, lot_number: '70 (T)', section: 1, size: 1
 lot71 = FactoryBot.create(:lot, lot_number: '71 (T)', section: 1, size: 1)
 property_977 = FactoryBot.create(:property, lots: [lot70, lot71], street_number: '977', street_name: 'Waterview Dr (TEST)')
 
-jr = FactoryBot.create(:resident, last_name: 'Rainwater', first_name: 'Jim', email_address: 'jim@example.com', properties: [property_977]).tap do |resident|
+pp = FactoryBot.create(:resident, last_name: 'PartialPayment', first_name: 'Peter', email_address: 'ppp@example.com', properties: [property_977]).tap do |resident|
   resident.residencies.first.update(
     resident_status: :deed_holder,
     verified_on: 1.day.ago)
@@ -87,12 +87,12 @@ end
 
 
 # Comments
-FactoryBot.create(:comment, commentable: mms)
+Resident.all.each {|r| FactoryBot.create_list(:comment, rand(5), commentable: r) }
 
 # Shares
 FactoryBot.create(:share_transaction, :purchase, quantity: 10, residency: mms.residencies.first)
-FactoryBot.create(:share_transaction, :purchase, quantity: 20, residency: jr.residencies.first)
-FactoryBot.create(:share_transaction, :transfer, quantity: 10, from_residency: jr.residencies.first, residency: cbs.residencies.first)
+FactoryBot.create(:share_transaction, :purchase, quantity: 20, residency: pp.residencies.first)
+FactoryBot.create(:share_transaction, :transfer, quantity: 10, from_residency: pp.residencies.first, residency: cbs.residencies.first)
 
 # Vehicles
 Resident.lot_fees_paid.each {|r| FactoryBot.create(:vehicle, resident: r) }