+++ CHANGELOG.md
@@ -16,6 +16,9 @@ In place of release version numbers, we organize via deploys to Production (by D
   - Index views: indicate resident_status & x of y models
     - Using new residencies/_property & _resident icon list views
   - Property#show: lists Lot_fee_paid? (as well as date) 
+- Seed
+  - Rename JimR to "Peter PartialPayment"
+  - Add (random) comments
 
 ## 2022-05-20: Filter by Scopes